[agent] $MAC_ADDRESS can be bridge
In many cases it is necessary to use bridge interface with ip interface
for management (for example Tinc interface, adhoc0 and eth0).
Though, it can be used as normal interface.
Also in some cases bridge interface can use its own mac address.
So it is better to exclude only loopback and tap interfaces.

References and closes #29.