[fix/change!] List options are now overwritten #82 #89
When using list options the old behavior was to merge
the options present in the controller with the ones
present locally, which made removing list options not
possible without hacky workarounds.

This change fixes the problem by ensuring list options
present in the controller overwrite the local ones.

Closes #82

It also solves issue https://github.com/openwisp/openwisp-config/issues/89
because list options do not go through the process of being converted
to a set and then back to a list (which shuffles the order).

Closes #89