[change] Remove unneeded ssl package implementation definitions
Which ssl implementation is used by the openwisp-config agent is decided
in the curl package. The openwisp-agent is not compiled servivce.
Therefore, the division into different packages is not
necessary. This commit removes all ssl package versions. There is now
only one openwisp-config package.

Other PRs related to this work:

- https://github.com/openwrt/luci/pull/5471
- https://github.com/openwisp/openwrt-openwisp-monitoring/pull/78

Signed-off-by: Florian Eckert <fe@dev.tdt.de>